requirements.txt
google-api-python-client
pandas
python-dotenv
psycopg2-binary
apache-airflow